Create pre_build script
Some notes. I check if Python 3.5.3 is already installed because we
don't want to install it everytime we deploy. Also, given that
virtualenv is already installed with python3 I just use it, without
check if the command exists. Also, cleaning up the $OPENSHIFT_TMP_DIR
is needed only if it contains something

Signed-off-by: Antonio Ossa <aaossa@uc.cl>