Create start script
This scripts will launch our application. At the beginning of the file
I created some useful variables that will be given to Gunicorn later.
I decided to save the pidfile in the $OPENSHIFT_DATA_DIR to read it
later in the stop script. First of all I try to stop the application
(just to avoid any problem with the port). Before run the Gunicorn
server I activate the virtual environment to be able to use the
dependencies installed in the build script. At the end of the file,
gunicorn use the app object from our demo file with the options shown

Signed-off-by: Antonio Ossa <aaossa@uc.cl>